Add comm_open/clode/msg message handling in kernel
The reason for the IoPubMessageSender is that the comm messages don't
use the normal request/response structure, but they might produce 0 or more
messages for each request, depending on the widget and the message.

Initially I wanted them to return an array (or range) of messages, but sometimes
messages are produced deep inside a widget. That makes a range version un-
feasible. The delegate version I implemented beats the array one since it
immediately sends messages and doesn't require a temp array buffer.